Added the ability for the user to choose whether Round Robin or CFS will be used as the default scheduler at compile-time.
Options are CONFIG_SCHED_NORMAL and CONFIG_SCHED_RR

I should probably put in a sanity check so that they aren't both
defined, but oh well.

Also added in a better README, listing the components of this patchset,
along with links to all of them and a little legal stuff.